feat: implement package browsing controllers, review management, and global Jest test infrastructure.

+++ src/core/Helpers/validation.js
@@ -390,7 +390,7 @@ export const schemas = {
 
     // Review Schemas
     submitReview: z.object({
-        rating: z.coerce.number().int().min(1).max(5),
+        rating: z.coerce.number().min(1).max(5),
         comment: z.string().max(1000).optional()
     })
 };
